FAI-317: minor changes
