+++ packages/pmml-editor/src/editor/components/DataDictionary/ConstraintsEdit/ConstraintsEdit.tsx
@@ -1,90 +1,64 @@
-import * as React from "react";
-import { useEffect, useState } from "react";
-import { v4 as uuid } from "uuid";
-import { omit, isEqual } from "lodash";
-import {
-  Button,
-  Card,
-  CardBody,
-  CardTitle,
-  Form,
-  FormGroup,
-  Select,
-  SelectOption,
-  SelectVariant,
-  Split,
-  SplitItem,
-  Stack,
-  StackItem,
-  Text,
-  TextContent,
-  TextVariants
-} from "@patternfly/react-core";
-import { ArrowAltCircleLeftIcon } from "@patternfly/react-icons";
-import ConstraintsEnumEdit from "../ConstraintsEnumEdit/ConstraintsEnumEdit";
-import { DDDataField, EnumConstraint, RangeConstraint } from "../DataDictionaryContainer/DataDictionaryContainer";
-import { Validated } from "../../../types";
-import "./ConstraintsEdit.scss";
-import ConstraintsRangeEdit from "../ConstraintsRangeEdit/ConstraintsRangeEdit";
+/*
+ * Copyright 2020 Red Hat, Inc. and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 
-interface ConstraintsEditProps {
+import * as React from "react";
+import { Card, CardBody, CardTitle, FormGroup, Select, SelectOption, SelectVariant } from "@patternfly/react-core";
+
+import { DDDataField, RangeConstraint } from "../DataDictionaryContainer/DataDictionaryContainer";
+import ConstraintsRangeEdit from "../ConstraintsRangeEdit/ConstraintsRangeEdit";
+import ConstraintsEnumEdit from "../ConstraintsEnumEdit/ConstraintsEnumEdit";
+import { useEffect, useState } from "react";
+import { Validated } from "../../../types";
+
+interface ConstraintsProps {
   dataType: DDDataField;
-  onSave: (payload: DDDataField) => void;
-  onClose: () => void;
+  onSave: (payload: Partial<DDDataField>) => void;
 }
 
-const ConstraintsEdit = (props: ConstraintsEditProps) => {
-  const { dataType, onSave, onClose } = props;
-  const [dataTypeState, setDataTypeState] = useState(dataType);
+const Constraints = (props: ConstraintsProps) => {
+  const { dataType, onSave } = props;
   const [constraintType, setConstraintType] = useState<string>(dataType.constraints?.type ?? "");
-  const [typeIsOpen, setTypeIsOpen] = useState(false);
+  const [typeSelectIsOpen, setTypeSelectIsOpen] = useState(false);
   const constraintsTypes = [{ value: "", isPlaceholder: true }, { value: "Range" }, { value: "Enumeration" }];
-  const [range, setRange] = useState<RangeConstraint[] | undefined>(
+  const [ranges, setRanges] = useState<RangeConstraint[] | undefined>(
     dataType.constraints?.type === "Range" ? dataType.constraints.value : undefined
   );
-  // const [enums, setEnums] = useState(
-  //   dataType.constraints?.type === "Enumeration" ? dataType.constraints.value : [{ value: "", id: uuid() }]
-  // );
+  const [enums, setEnums] = useState(
+    dataType.constraints?.type === "Enumeration" ? dataType.constraints.value : undefined
+  );
   const [validation, setValidation] = useState<FormValidation>({
     form: "default",
     fields: { start: "default", end: "default", enums: "default" }
   });
 
-  useEffect(() => {
-    if (!isEqual(omit(dataType, "constraints"), omit(dataTypeState, "constraints"))) {
-      onClose();
-    } else {
-      setDataTypeState(dataType);
-    }
-  }, [dataType, dataTypeState]);
-
-  useEffect(() => {
-    setConstraintType(dataType.constraints?.type ?? "");
-    if (dataType.constraints?.type === "Range") {
-      setRange(dataType.constraints.value);
-    }
-    // if (dataType.constraints?.type === "Enumeration") {
-    //   setEnums(dataType.constraints.value);
-    // }
-  }, [dataType.constraints, setConstraintType, setRange]);
-
   const handleTypeChange = (event: React.MouseEvent | React.ChangeEvent, value: string) => {
     if (value !== constraintType) {
       setConstraintType(value);
-      setValidation({ form: "default", fields: { start: "default", end: "default", enums: "default" } });
       if (value === "Range") {
         onSave({
-          ...dataType,
           constraints: {
             type: "Range",
             value: [
               {
                 start: {
-                  value: "1",
+                  value: "",
                   included: true
                 },
                 end: {
-                  value: "10",
+                  value: "",
                   included: true
                 }
               }
@@ -92,187 +66,183 @@ const ConstraintsEdit = (props: ConstraintsEditProps) => {
           }
         });
       }
+      if (value === "Enumeration") {
+        onSave({
+          constraints: {
+            type: "Enumeration",
+            value: [""]
+          }
+        });
+      }
     }
-    setTypeIsOpen(false);
+    setTypeSelectIsOpen(false);
+  };
+  const handleTypeToggle = () => {
+    setTypeSelectIsOpen(!typeSelectIsOpen);
   };
 
-  const handleTypeOpening = () => {
-    setTypeIsOpen(!typeIsOpen);
-  };
-
-  const handleRangeChange = (rangeValue: RangeConstraint) => {
+  const handleRangeSave = (updatedRanges: RangeConstraint[]) => {
     onSave({
-      ...dataType,
       constraints: {
         type: "Range",
-        value: [rangeValue]
+        value: updatedRanges
       }
     });
   };
 
-  // const handleEnumsChange = (value: string, index: number) => {
-  //   const updatedEnums = [...enums];
-  //   updatedEnums[index].value = value;
-  //   setEnums(updatedEnums);
-  // };
-  //
-  // const handleEnumsDelete = (index: number) => {
-  //   const updatedEnums = [...enums];
-  //   updatedEnums.splice(index, 1);
-  //   setEnums(updatedEnums);
-  // };
-  //
-  // const handleAddEnum = () => {
-  //   const updatedEnums = [...enums];
-  //   updatedEnums.push({ value: "", id: uuid() });
-  //   setEnums(updatedEnums);
-  // };
-  //
-  // const handleEnumSort = (oldIndex: number, newIndex: number) => {
-  //   const newOrder = reorderArray(enums, oldIndex, newIndex);
-  //   setEnums(newOrder);
-  // };
-
-  const validateConstraints = () => {
-    const isValid = { ...validation };
-    isValid.form = "success";
-    // if (constraintType === "Range") {
-    //   isValid.fields.start = range?.start.value.trim().length === 0 ? "error" : "success";
-    //   isValid.fields.end = range?.end.value.trim().length === 0 ? "error" : "success";
-    // }
-    // if (constraintType === "Enumeration") {
-    //   const oneEnum = enums.find(item => item.value.trim().length > 0);
-    //   isValid.fields.enums = oneEnum === undefined ? "error" : "success";
-    // }
-    for (const fieldsKey in isValid.fields) {
-      if (isValid.fields[fieldsKey] === "error") {
-        isValid.form = "error";
+  const handleRangeAdd = () => {
+    const updatedRanges = [...ranges];
+    updatedRanges.push({
+      start: {
+        value: "",
+        included: true
+      },
+      end: {
+        value: "",
+        included: true
       }
-    }
-    setValidation(isValid);
-    return isValid.form;
+    });
+    onSave({
+      constraints: {
+        type: "Range",
+        value: updatedRanges
+      }
+    });
   };
 
-  const handleSubmit = (event: React.FormEvent) => {
-    event.preventDefault();
-    if (validateConstraints() === "success") {
-      onClose();
+  const handleRangeDelete = (index: number) => {
+    const updatedRanges = [...ranges];
+    updatedRanges.splice(index, 1);
+    onSave({
+      constraints: {
+        type: "Range",
+        value: updatedRanges
+      }
+    });
+  };
+
+  const handleEnumsChange = (value: string, index: number) => {
+    const updatedEnums = [...enums];
+    updatedEnums[index] = value;
+    onSave({
+      constraints: {
+        type: "Enumeration",
+        value: updatedEnums
+      }
+    });
+  };
+
+  const handleEnumsDelete = (index: number) => {
+    const updatedEnums = [...enums];
+    updatedEnums.splice(index, 1);
+    onSave({
+      constraints: {
+        type: "Enumeration",
+        value: updatedEnums
+      }
+    });
+  };
+
+  const handleAddEnum = () => {
+    const updatedEnums = [...enums, ""];
+    onSave({
+      constraints: {
+        type: "Enumeration",
+        value: updatedEnums
+      }
+    });
+  };
+
+  const handleEnumSort = (oldIndex: number, newIndex: number) => {
+    if (enums) {
+      const updatedEnums = reorderArray(enums, oldIndex, newIndex);
+      onSave({
+        constraints: {
+          type: "Enumeration",
+          value: updatedEnums
+        }
+      });
     }
   };
 
-  const clearConstraints = () => {
-    const updatedDataType = { ...dataType };
-    delete updatedDataType.constraints;
-    onSave(updatedDataType);
-    setValidation({ form: "default", fields: {} });
-  };
+  useEffect(() => {
+    setConstraintType(dataType.constraints?.type ?? "");
+    if (dataType.constraints?.type === "Range") {
+      setRanges(dataType.constraints.value);
+    }
+    if (dataType.constraints?.type === "Enumeration") {
+      setEnums(dataType.constraints.value);
+    }
+  }, [dataType.constraints]);
 
   return (
-    <section>
-      <Stack hasGutter={true}>
-        <StackItem>
-          <TextContent>
-            <Text component={TextVariants.h3}>
-              <Button variant="link" isInline={true} onClick={event => handleSubmit(event)}>
-                {dataType.name}
-              </Button>
-              &nbsp;/&nbsp;Constraints
-            </Text>
-          </TextContent>
-        </StackItem>
-        <StackItem>
-          <section className="constraints__form">
-            <Form onSubmit={handleSubmit} autoComplete="off">
-              <FormGroup
-                fieldId="constraints-type"
-                label="Constraints Type"
-                helperText="Select the type of constraint and then fill in the required fields."
-              >
-                {/*PF 2020.08 has a bug setting width of a Select*/}
-                <div style={{ width: 300 }}>
-                  <Select
-                    id="constraints-type"
-                    variant={SelectVariant.single}
-                    aria-label="Select Constraint Type"
-                    onToggle={handleTypeOpening}
-                    onSelect={handleTypeChange}
-                    selections={constraintType}
-                    isOpen={typeIsOpen}
-                    placeholderText={"Select a type"}
-                  >
-                    {constraintsTypes.map((item, index) => (
-                      <SelectOption key={index} value={item.value}>
-                        {item.isPlaceholder ? "Select a type" : item.value}
-                      </SelectOption>
-                    ))}
-                  </Select>
-                </div>
-              </FormGroup>
-              {/*range constraints editing to be moved to a dedicated component*/}
-              {constraintType === "Range" && range !== undefined && (
-                <Card isCompact={true}>
-                  <CardTitle>Range Constraint</CardTitle>
-                  <CardBody>
-                    {/*<ConstraintsRangeEdit*/}
-                    {/*  range={range}*/}
-                    {/*  onChange={handleRangeChange}*/}
-                    {/*  typeOfData={dataType.type}*/}
-                    {/*  validation={validation}*/}
-                    {/*/>*/}
-                  </CardBody>
-                </Card>
-              )}
-              {constraintType === "Enumeration" && (
-                <Card isCompact={true}>
-                  <CardTitle>Enumerations List</CardTitle>
-                  <CardBody>
-                    {/*<ConstraintsEnumEdit*/}
-                    {/*  enumerations={enums}*/}
-                    {/*  onChange={handleEnumsChange}*/}
-                    {/*  onDelete={handleEnumsDelete}*/}
-                    {/*  onAdd={handleAddEnum}*/}
-                    {/*  onSort={handleEnumSort}*/}
-                    {/*  validation={validation}*/}
-                    {/*/>*/}
-                  </CardBody>
-                </Card>
-              )}
-            </Form>
-          </section>
-        </StackItem>
-        <StackItem>
-          <Split hasGutter={true}>
-            <SplitItem>
-              <Button
-                variant="primary"
-                type="button"
-                onClick={handleSubmit}
-                icon={<ArrowAltCircleLeftIcon />}
-                iconPosition="left"
-              >
-                Done
-              </Button>
-            </SplitItem>
-            <SplitItem>
-              <Button variant="secondary" type="button" onClick={clearConstraints}>
-                Clear Constraints
-              </Button>
-            </SplitItem>
-          </Split>
-        </StackItem>
-      </Stack>
+    <section className="constraints__form">
+      <FormGroup
+        fieldId="constraints-type"
+        label="Constraints Type"
+        helperText="Select the type of constraint and then fill in the required fields."
+      >
+        {/*PF 2020.08 has a bug setting width of a Select*/}
+        <div style={{ width: 300 }}>
+          <Select
+            id="constraints-type"
+            variant={SelectVariant.single}
+            aria-label="Select Constraint Type"
+            onToggle={handleTypeToggle}
+            onSelect={handleTypeChange}
+            selections={constraintType}
+            isOpen={typeSelectIsOpen}
+            placeholderText={"Select a type"}
+          >
+            {constraintsTypes.map((item, index) => (
+              <SelectOption key={index} value={item.value}>
+                {item.isPlaceholder ? "Select a type" : item.value}
+              </SelectOption>
+            ))}
+          </Select>
+        </div>
+      </FormGroup>
+      {constraintType === "Range" && ranges !== undefined && (
+        <Card isCompact={true} style={{ margin: "1em 0" }}>
+          <CardTitle>Range Constraint</CardTitle>
+          <CardBody>
+            <ConstraintsRangeEdit
+              ranges={ranges}
+              onAdd={handleRangeAdd}
+              onChange={handleRangeSave}
+              onDelete={handleRangeDelete}
+              validation={validation}
+            />
+          </CardBody>
+        </Card>
+      )}
+      {constraintType === "Enumeration" && enums !== undefined && (
+        <Card isCompact={true} style={{ margin: "1em 0" }}>
+          <CardTitle>Enumerations List</CardTitle>
+          <CardBody>
+            <ConstraintsEnumEdit
+              enumerations={enums}
+              onChange={handleEnumsChange}
+              onDelete={handleEnumsDelete}
+              onAdd={handleAddEnum}
+              onSort={handleEnumSort}
+              validation={validation}
+            />
+          </CardBody>
+        </Card>
+      )}
     </section>
   );
 };
 
-export default ConstraintsEdit;
+export default Constraints;
 
 export interface FormValidation {
   form: Validated;
   fields: { [key: string]: Validated };
 }
 
-const reorderArray = (list: EnumConstraint[], startIndex: number, endIndex: number) => {
+const reorderArray = <T extends unknown>(list: T[], startIndex: number, endIndex: number) => {
   const result = [...list];
   const [removed] = result.splice(startIndex, 1);
   result.splice(endIndex, 0, removed);
