+++ packages/pmml-editor/src/editor/components/DataDictionary/DataTypeItem/DataTypeItem.scss
@@ -6,7 +6,8 @@
   &__name {
     font-weight: 700;
     word-break: keep-all;
-    line-height: 2.1875rem;
+    display: block;
+    margin-top: 5px;
   }
   &__type-label {
     text-transform: capitalize;
